[gDM] Do not allow capturing Chromium windows audio in Picker UI
This CL changes the behavior of the getDisplayMedia picker UI for
application audio capture scenarios where a Chromium window is selected
to be shared. In this scenario, Chromium audio should not be shared and
the picker UI audio share toggle should be disabled until a non-Chromium
window is selected.

For a visual representation of this behavior, please refer to the GIF in
the "Privacy considerations" section of this feature's design doc:
https://docs.google.com/document/d/1lvHAeM5uYgM6HSdfU9lel2VhSnbNRsdZE5Ph4rBiXOY/edit?tab=t.0#heading=h.qjpzx8d6k7t5

Bug: 40947205
Change-Id: I9ae87cad23611cf6554c12d8257dffc86f536ff3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7169457
Reviewed-by: Johannes Kron <kron@chromium.org>
Commit-Queue: Gabriel Brito <gabrielbrito@microsoft.com>
Reviewed-by: Tove Petersson <tovep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550143}