ServiceWorker: Add metrics to validate SHA256 checksum Service Worker.
This change introduces a mechanism to verify the integrity of Service
Worker resources by calculating a SHA256 hash of the resource data as
it's read and comparing it against a checksum stored in the resource
metadata.

The result of the checksum comparison is recorded via a new UMA
histogram, "ServiceWorker.ResourceChecksumMatch".

Bug: 458516159
Change-Id: Iefe4ba2923bf36cc850a0fcdf7046a3246225688
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7130007
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Auto-Submit: Eriko Kurimoto <elkurin@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550223}