Allow remote debugging server to fall back to an available port
When starting the remote debugging server, first attempt to bind to the
default port (9222). If that port is in use, try to find any available
port on both IPv4 and IPv6 addresses. The free IPv4 address is preferred
to avoid a confusion between 127.0.0.1:9222 and [::1]:9222 being used at
the same time.

Bug: 460665929
Change-Id: Ibb85b51cfd4b61ad341b46bf1498a465c5870a73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7200659
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550319}