[permissions] Refactor bool into enum for GeolocationAccuracy
This is a pure refactoring, with no functional changes.

This CL exposes the GeolocationAccuracy c++ enum to java and gets rid
of the java enum LocationAccuracy. It also refactors the c++/java
bindings and the GeolocationPromptOptions struct to use the enum
instead of a boolean.

Change-Id: I4dc0b362f564466abfff1dd74a12590c9499401e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201450
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550342}