[SyncTests] Update SyncServiceImplHarness callers
This CL updates all callers of
SyncServiceImplHarness::EnableSyncForType() and DisableSyncForType() to
use EnableSelectableType()/DisableSelectableType() instead, and removes
the former.

DisableSyncForType() and DisableSelectableType() do exactly the same
thing, so this part is a no-op.
For EnableSyncForType(), there is a difference: It also turns on
Sync-the-feature, if it wasn't already enabled. It turns out that none
of its callers actually depended on this behavior, so the simpler
EnableSelectedType() (which only turns on the type) does the job too.

Bug: 353425612
Change-Id: I368651ff796842e3218e14a547acc9ed6a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7198367
Reviewed-by: Mahmoud Rashad <mmrashad@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550360}