[BNPL][Clank] Refactor error screen dismissal to use onDismissed
This CL refactors the error screen's "OK" button to use the
`onDismissed` method instead of its own `onErrorOkPressed` handler.

This change centralizes the dismissal logic for the TTF bottom sheet.
This is a prerequisite for a follow-up CL that will allow the bottom
sheet to be re-shown after a user dismisses it during a BNPL flow.

Bug: 430575808
Change-Id: I45ae4bbbb8ff404486ce4b44956000b306cd9fa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7204742
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Fernando Ramirez <ferny@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1550368}