[PEPC] Propagate PEPC descriptor on Android and decouple cleanup
Changes in this CL:
- First, it ensures that the `embedded_permission_request_descriptor`
is correctly propagated when handling geolocation permission requests
on Android. Previously, this data was lost when overriding the
`PermissionRequestData` in the Android delegate, causing the context
to lose track of whether the request was PEPC-initiated.
- Second, it modifies `PermissionContextBase` to decouple the cleaning
up of PEPC and regular request. PEPC's new cleaning up callback ensures
the correct order between `MaybeUpdateCachedHasDevicePermission` and
triggering the decision callback.

Fixed: 463327697, 461402995
Change-Id: Ie93099638c149986b2dbef064d9f2014a73444ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7198102
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Commit-Queue: Thomas Nguyen <tungnh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550369}