[iOS] Remove table view from infobarmodal
The infobar modals are using a TableView to display their content, but
there is no reason for it.

This CL replaces it with a stack view and ContentConfigurations.

Before:
https://screenshot.googleplex.com/AEME3ZxkBikVs2S
https://screenshot.googleplex.com/DNA9TVpovWwxUEv
https://screenshot.googleplex.com/BqAvuAZEke5tttV

After:
https://screenshot.googleplex.com/A53TS9G48dMZgGC
https://screenshot.googleplex.com/93rR3SFEdfuMU89
https://screenshot.googleplex.com/8LZTQCBnXGm2HMu

BYPASS_RECITATION_REASON=This is a rewrite of existing tests. Yes it was done with the help of Gemini but it is really how I would have done it manually. I don't know how to modify it to avoid the recitation checker.

Bypass-Check-License: file moved
Bug: 443034511
Change-Id: Iaa849ee33e09c3060013cf698e9a771f3ba36fb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7184360
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1550386}