Add Target.getDevToolsTarget CDP command.
This command allows retrieving the targetId of the DevTools page target associated with a given targetId, if a DevTools window is currently open for it. The implementation is delegated to the DevToolsManagerDelegate, with Chrome providing the logic to find the DevTools window for a given agent host. A new browser test is added to verify the functionality.

Bug: 463637554
Change-Id: I8c4883c01815af69e5a1d3ae59e7c57f31703d36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7205468
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550406}