Store layer association in the RuleSet (@page)
We currently set the CascadeLayer association directly on the rule
as a side effect of RuleSet building. This approach does not work,
as it may lead to rules that are associated with non-existent layers
(see Issue 463698792).

This CL (first in a chain of many) introduces a CascadeLayered<T>
template class, and stores a vector of CascadeLayered<StyleRulePage>
in the RuleSet rather than a plain vector of StyleRulePage.

We seem to be missing tests for layered @page entirely; this CL
adds some.

Fixes DCHECK involving unconnected RuleSets and rebuilding of
the cascade layer map (CreateUnconnectedRuleSets_LayeredPageRule,
Issue 428664521).

I also discovered that we're not sorting @page rules correctly
across RuleSets, but that's not addressed here (Issue 463591206).

This is expected to slightly increase memory usage: although a rule
is typically only stored in a single RuleSet, it usually also exists
in at least one other container on ScopedStyleResolver. The assumption
is that we have so few of these rules that the memory impact
is acceptable.

At the end of this CL chain, I plan to remove the "int" version
of CompareLayerOrder, and just use std::weak_ordering everywhere.

Issue: 463698792, 428664521, 463591206
Change-Id: If1237e53df78f2421b5513fd6e160efd3674d096
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7204862
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550437}