net/http: Upgrade SHA-1 to SHA-256 in HttpCache
SHA-1 is cryptographically broken and deprecated. This change upgrades
the in-memory no-store key tracking to use SHA-256.

SHA-1 is currently only used for storing a list of keys that we
shouldn't store in the HTTP cache. This capability is used only in
http_cache_transaction.cc, and is gated behind a feature flag
`kAvoidEntryCreationForNoStore`.

Bug: 430344933
Change-Id: I51b639e76fb7b84e3dccb791ac801f19767ce26f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7170337
Commit-Queue: Antonio Alphonse <alphonsea@google.com>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550510}