Add rp_id as a field in CredentialUIEntry
For credential export, there is a need to match a user-selected
CredentialUIEntry to a passkey proto in Sync, in order to fetch some
data that should not be stored in CredentialUIEntry itself (e.g. private
key). That requires a pair of (credential_id, rp_id) and for that
purpose rp_id is added as a field in CredentialUIEntry in this CL.

https://crbug.com/463623303 was added as a follow-up improvement to make
passkey-only fields of CredentialUIEntry an optional struct.

Bug: 463342373
Change-Id: Ia2c5531c572a3c0a617eeb065018672e01db333e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7203165
Commit-Queue: Rafał Godlewski <rgod@google.com>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1550530}