Store layer association in the RuleSet (@keyframes)
We currently set the CascadeLayer association directly on the rule
as a side effect of RuleSet building. This approach does not work,
as it may lead to rules that are associated with non-existent layers
(see Issue 463698792); we need to store the layer association next to
the rule in the RuleSet instead.

Fixes DCHECK involving unconnected RuleSets and rebuilding of
the cascade layer map (CreateUnconnectedRuleSets_LayeredKeyframes,
Issue 428664521).

This is expected to slightly increase memory usage: although a rule
is typically only stored in a single RuleSet, it usually also exists
in at least one other container on ScopedStyleResolver. The assumption
is that we have so few of these rules that the memory impact
is acceptable.

Issue: 463698792, 428664521
Change-Id: I22754732a3862802f185cfd5fa0d12286d6da74c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7206226
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550536}