Updates the VariationsIdsProvider to take a base::Clock.
This CL updates variations::VariationsIdsProvider to take
a base::Clock instance (by unique_ptr) in its constructor
instead of using a Callback as its time source.

The places where the VariationsIdsCreator is initialized
are updated to provide it with a base::DefaultClock. A
future CL will update the Chrome's implementation of
content::MainDelegate to provide a VariationsNetworkClock
when constructing the VariationsIdsProvider.

See also:

* https://chromium-review.googlesource.com/c/chromium/src/+/7200463

Bug: 442849530
Change-Id: I7b3e287f10f345a401617604fcdc7109a8b27771
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7199040
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550561}