[iOS] Handle multiple concurrent passkey requests
This CL adds a request ID to every passkey request so that multiple
concurrent passkey requests can happen safely.

A map of deferred promises was added to replace the single promise
that was used before. Resolving a request or deferring to the renderer
now requires a request id. Once the promise is resolved or rejected,
the promise is removed from the map.

Bug: 460485333
Change-Id: I5ae83d386959610db5edc1a86d9deeeac29d7907
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7205263
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: Rafał Godlewski <rgod@google.com>
Reviewed-by: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1550563}