[LNA] check LNA permission pre-connection for definite-LNA request
Request the LNA permission even before the connection is established
when we know that the request will result in an LNA check. This is true
for 3 types of hostnames:

* IP literals
* localhost (and .localhost domains)
* .local domains

If the host resolves and the host matches one of the 3 above cases, then
a LNA prompt will be triggered even if the connection to the host is not
established.

This is being done to give developers a deterministic way to invoke an
LNA permission check even if there is nothing listening at the host.
Ideally the LNA check would be done after hostname resolution, which
would achieve the same behaviour, but this is a harder task and so this
limited-in-scope method for deterministically invoking an LNA permission
check.

Side effects of this CL:

* changed 2 test classes that were using localhost to be cross-origin
from 127.0.0.1, which was triggering the LNA permission (and not
counting as public because no ip-address-space-override is applied to
localhost domain)

* removed Security.PrivateNetworkAccess.PrivateIpInferrable histogram (CL would have changed how often it was logged, and this histogram was used to measure impact pre-launch; now that we've launched it is no longer needed).

DISABLE_SPELLCHECKER
OBSOLETE_HISTOGRAMSecurity.PrivateNetworkAccess.PrivateIpInferrable=No longer needed as LNA launched.

Bug: 461746665
Change-Id: Ica0c697d730ad3375f16eb4f553865f6a305acff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7167750
Reviewed-by: mmenke <mmenke@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Hubert Chao <hchao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550573}