[Contextual Tasks] Style side panel toolbar
This patch styles the top toolbar to be closer to the spec. A
placeholder image for the Google icon is in place temporarily. The
other icons need to be finalized as well - this patch just focuses
on the layout.

Screenshot: https://screenshot.googleplex.com/BvfHeVLHZx9VjDq

Bug: 461535004
Change-Id: I50edfcb338f04698b25fbb2ae1f4345a358d4242
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7204296
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550577}