Move SearchboxHandler to c/b/u/w/cr_components/searchbox/
Per this comment thread,
https://chromium-review.googlesource.com/c/chromium/src/+/7139182/comment/d7f16179_ab39e098/,
to allow the ComposeboxHandler, which is a SearchboxHandler, to have a
default config for all embedders, the files need to be moved to
cr_components. This CL is the first step in that process. This CL:
  - Moves searchbox_handler.h/cc to cr_components
  - Moves searchbox_omnibox_client, contextual_search_type_converters, and contextual_searchbox_handler, since those will be needed by ComposeboxHandler
  - Updated references to the new location
  - Fixed build file errors

Note: Since the current searchbox directory has a circular dependency
exception, I included the same for the new directory. If not, there are
10+ includes that are not allowed due to circular dependencies.

Moving ComposeboxHandler will happen in a separate CL.

Bug: b:459833598
Bypass-Check-License: Moved to a different location, not a new file.
Change-Id: I310b8a3c0caa771e5ec923b0991541f67741117c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7200957
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1550596}