[SyncTests] Modernize SingleClientWalletSyncTest
...by parameterizing them to also test the upcoming transport-only mode.

Generally no logical changes are required for this type, with the
exception of tests that exercise the sync-paused state (replaced by the
sign-in pending state). However, the helper functions need to be updated
significantly to allow toggling which underlying store to use, which
requires updating some additional files too (trivial changes).

In one test, a secondary test expectation failed to hold true in
transport mode. Fixing this seems complicated, so in this patch the
expectation is relaxed with a TODO to reintroduce it in follow-up
patches.

Bug: 353425612
Change-Id: I4e1e63280ef5b9fd6c8dc767e3554a1c37fcb681
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201860
Reviewed-by: Mahmoud Rashad <mmrashad@google.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550598}